fix for deserialisation error

+++ KcY-SeeItemValue/itemValue/jsonClass.cs
@@ -4,7 +4,7 @@
     {
         public double multiplier { get; set; }
         public int price { get; set; }
-        public int originalMax { get; set; }
+        public double originalMax { get; set; }
         public string traderName { get; set; }
     }
 }
